made solver a bit more efficient
found out that the program does not have very good memory. for example, if i put in the word audio, with the verdict as bybbb, in the later guesses, it would still give me words like could and would, even though the verdict clearly says that no, b and o are not in the word.